refactor(output): restore OutputData as canonical creator implementation
Restores OutputData as the canonical creator. Adapts DefaultOutputDataCreator as a thin wrapper around OutputData, removing the need for a global instance and minimizing the risks of circular deps in future.
