+++ src/model/OutputData.ts
@@ -1,17 +1,22 @@
-import { bytesToHex, hexToBytes } from '@noble/hashes/utils.js';
+import { bytesToHex, hexToBytes, randomBytes } from '@noble/hashes/utils.js';
 
 import {
+  blindMessage,
   constructUnblindedSignature,
+  deriveBlindingFactor,
+  deriveP2BKBlindedPubkeys,
+  deriveSecret,
+  normalizeP2PKOptions,
   pointFromHex,
   verifyDLEQProof,
   type DLEQ,
   type BlindSignature,
   type P2PKOptions,
 } from '../crypto';
-import { numberToHexPadded64 } from '../utils';
+import { Bytes, numberToHexPadded64, splitAmount } from '../utils';
 
 import { Amount, type AmountLike } from './Amount';
-import { defaultOutputDataCreator } from './OutputDataCreator';
+import { BlindedMessage } from './BlindedMessage';
 import {
   type HasKeysetKeys,
   type Proof,
@@ -148,19 +153,124 @@ export class OutputData implements OutputDataLike {
     keyset: HasKeysetKeys,
     customSplit?: AmountLike[],
   ): OutputData[] {
-    return defaultOutputDataCreator.createP2PKData(p2pk, amount, keyset, customSplit);
+    const amounts = splitAmount(amount, keyset.keys, customSplit);
+    return amounts.map((a) => this.createSingleP2PKData(p2pk, a, keyset.id));
   }
 
-  static createSingleP2PKData(p2pk: P2PKOptions, amount: AmountLike, keysetId: string) {
-    return defaultOutputDataCreator.createSingleP2PKData(p2pk, amount, keysetId);
+  static createSingleP2PKData(p2pk: P2PKOptions, amount: AmountLike, keysetId: string): OutputData {
+    const amountValue = Amount.from(amount);
+    const normalized = normalizeP2PKOptions(p2pk);
+    const lockKeys = Array.isArray(normalized.pubkey) ? normalized.pubkey : [normalized.pubkey];
+    const refundKeys = normalized.refundKeys ?? [];
+    const reqLock = normalized.requiredSignatures ?? 1;
+    const reqRefund = normalized.requiredRefundSignatures ?? 1;
+
+    // Init vars
+    const hashlock = normalized.hashlock;
+    const isHTLC = typeof hashlock === 'string' && hashlock.length > 0;
+    let data = isHTLC ? hashlock : lockKeys[0];
+    let pubkeys = isHTLC ? lockKeys : lockKeys.slice(1);
+    let refund = refundKeys;
+
+    // Optional key blinding (P2BK)
+    let Ehex: string | undefined;
+    if (p2pk.blindKeys) {
+      const ordered = [...lockKeys, ...refundKeys];
+      const { blinded, Ehex: _E } = deriveP2BKBlindedPubkeys(ordered);
+      if (isHTLC) {
+        // hashlock is in data, all locking keys into pubkeys
+        pubkeys = blinded.slice(0, lockKeys.length);
+      } else {
+        // first locking key in data, rest into pubkeys
+        data = blinded[0];
+        pubkeys = blinded.slice(1, lockKeys.length);
+      }
+      refund = blinded.slice(lockKeys.length);
+      Ehex = _E;
+    }
+
+    // build P2PK Tags (NUT-11)
+    const tags: string[][] = [];
+
+    const ts = normalized.locktime ?? NaN;
+    if (Number.isSafeInteger(ts) && ts >= 0) {
+      tags.push(['locktime', String(ts)]);
+    }
+
+    if (pubkeys.length > 0) {
+      tags.push(['pubkeys', ...pubkeys]);
+      if (reqLock > 1) {
+        tags.push(['n_sigs', String(reqLock)]);
+      }
+    }
+
+    if (refund.length > 0) {
+      tags.push(['refund', ...refund]);
+      if (reqRefund > 1) {
+        tags.push(['n_sigs_refund', String(reqRefund)]);
+      }
+    }
+
+    if (normalized.sigFlag == 'SIG_ALL') {
+      tags.push(['sigflag', 'SIG_ALL']);
+    }
+
+    // Append additional tags if any
+    if (normalized.additionalTags?.length) {
+      const extraTags = normalized.additionalTags.map(([k, ...vals]) => {
+        assertValidTagKey(k); // Validate key
+        return [k, ...vals.map(String)]; // all to strings
+      });
+      tags.push(...extraTags);
+    }
+
+    // Construct secret
+    const kind = isHTLC ? 'HTLC' : 'P2PK';
+    const newSecret: [string, { nonce: string; data: string; tags: string[][] }] = [
+      kind,
+      {
+        nonce: bytesToHex(randomBytes(32)),
+        data,
+        tags,
+      },
+    ];
+    const parsed = JSON.stringify(newSecret);
+
+    // Check secret length, counting Unicode code points
+    // Same semantics as Nutshell python: len(str)
+    const charCount = [...parsed].length;
+    if (charCount > MAX_SECRET_LENGTH) {
+      throw new Error(`Secret too long (${charCount} characters), maximum is ${MAX_SECRET_LENGTH}`);
+    }
+
+    // blind the message
+    const secretBytes = new TextEncoder().encode(parsed);
+    const { r, B_ } = blindMessage(secretBytes);
+
+    // create OutputData
+    return new OutputData(
+      new BlindedMessage(amountValue, B_, keysetId).getSerializedBlindedMessage(),
+      r,
+      secretBytes,
+      Ehex,
+    );
   }
 
   static createRandomData(amount: AmountLike, keyset: HasKeysetKeys, customSplit?: AmountLike[]) {
-    return defaultOutputDataCreator.createRandomData(amount, keyset, customSplit);
+    const amounts = splitAmount(amount, keyset.keys, customSplit);
+    return amounts.map((a) => this.createSingleRandomData(a, keyset.id));
   }
 
-  static createSingleRandomData(amount: AmountLike, keysetId: string) {
-    return defaultOutputDataCreator.createSingleRandomData(amount, keysetId);
+  static createSingleRandomData(amount: AmountLike, keysetId: string): OutputData {
+    const amountValue = Amount.from(amount);
+    const randomHex = bytesToHex(randomBytes(32));
+    const secretBytes = new TextEncoder().encode(randomHex);
+    const { r, B_ } = blindMessage(secretBytes);
+    return new OutputData(
+      new BlindedMessage(amountValue, B_, keysetId).getSerializedBlindedMessage(),
+      r,
+      secretBytes,
+    );
   }
 
   static createDeterministicData(
@@ -170,12 +280,9 @@ export class OutputData implements OutputDataLike {
     keyset: HasKeysetKeys,
     customSplit?: AmountLike[],
   ): OutputData[] {
-    return defaultOutputDataCreator.createDeterministicData(
-      amount,
-      seed,
-      counter,
-      keyset,
-      customSplit,
+    const amounts = splitAmount(amount, keyset.keys, customSplit);
+    return amounts.map((a, i) =>
+      this.createSingleDeterministicData(a, seed, counter + i, keyset.id),
     );
   }
 
@@ -188,8 +295,20 @@ export class OutputData implements OutputDataLike {
     seed: Uint8Array,
     counter: number,
     keysetId: string,
-  ) {
-    return defaultOutputDataCreator.createSingleDeterministicData(amount, seed, counter, keysetId);
+  ): OutputData {
+    const amountValue = Amount.from(amount);
+    const secretBytes = deriveSecret(seed, keysetId, counter);
+    const secretBytesAsHex = bytesToHex(secretBytes);
+    const utf8SecretBytes = new TextEncoder().encode(secretBytesAsHex);
+    // Note: Bytes.toBigInt is used here so invalid values bubble up as throws
+    // for BIP32-style retry logic (caller increments counter and retries).
+    const deterministicR = Bytes.toBigInt(deriveBlindingFactor(seed, keysetId, counter));
+    const { r, B_ } = blindMessage(utf8SecretBytes, deterministicR);
+    return new OutputData(
+      new BlindedMessage(amountValue, B_, keysetId).getSerializedBlindedMessage(),
+      r,
+      utf8SecretBytes,
+    );
   }
 
   /**
